refactor(terminal): make the daemon the only session backend
Deletes the in-process fork. `TERMIO_TERMIOD` and every branch on it are
gone, along with `PTYProcess` itself, the `ptyProcesses` table, and the
flag-off placeholders and alerts that existed only to explain why a remote
session could not open.

Every capability the removed path carried has a counterpart the daemon has
been reporting since the session-facts work: the status tap becomes `E
status`, the foreground poll becomes `foreground_argv` on the roster, and
the exit handling becomes `session_exited` routed through the same shared
`sessionExit` policy. `closeConfirmationReason` reads the device's answer
directly now that there is no local answer to reconcile it with.

`reapStrayOrphans` survives as `StraySessionReaper`, and the reason it exists
survives with it. The owner it defends against is now `termiod` rather than
this app, but the leak has the same shape — a session whose owner died
without teardown re-parents to launchd, keeps its dead PTY, and idles forever
because agent TUIs swallow SIGHUP. A tombstone records that a session ended;
it does not reach the process, and the daemon that would have killed the
group is the thing that died.

`PTYForegroundTests` goes too. It covered the in-process live-member search,
whose counterpart in `proc.rs` has its own tests.

One path, local and remote alike: a socket for this Mac, SSH stdio for a VPS,
and the same framed protocol on both.
